mcp-servers/calendar/requirements.txt
fastmcp>=3.0.0
httpx>=0.27.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
pytz>=2024.0
